+++ CHANGES.md
@@ -2,6 +2,7 @@
 
 ## Unreleased
 - Added transformer for MongoDB CDC to CrateDB SQL conversion
+- Rename "Airrohr" decoder to "Sensor.Community"
 
 ## 2024/07/16 v0.0.1
 - Added decoders for Airrohr, Tasmota, and TTS/TTN from Kotori DAQ